Make a portable assertions for ruby core merging.
