+++ test/test_racc_command.rb
@@ -141,13 +141,13 @@ module Racc
     end
 
     def test_norule_y
-      assert_raise(Test::Unit::AssertionFailedError) {
+      assert_raise(MiniTest::Assertion) {
         assert_compile 'norule.y'
       }
     end
 
     def test_unterm_y
-      assert_raise(Test::Unit::AssertionFailedError) {
+      assert_raise(MiniTest::Assertion) {
         assert_compile 'unterm.y'
       }
     end
